md_converter/requirements.txt
# Core dependencies
markdown>=3.5.0
python-docx>=1.1.0
weasyprint>=60.0
Pillow>=10.0.0
pygments>=2.16.0

# Markdown extensions
markdown-it-py>=3.0.0
pymdown-extensions>=10.0

# Diagram processing
pydot>=2.0.0

# ASCII art processing
art>=6.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
PyYAML>=6.0
python-magic>=0.4.27
beautifulsoup4>=4.12.0
lxml>=4.9.0
